NP-45038 Add flow for creating registration from import candidate
