+++ src/pages/basic_data/BasicDataPage.tsx
@@ -24,7 +24,6 @@ import { UrlPathTemplate, getAdminInstitutionPath } from '../../utils/urlPaths';
 import { AdminCustomerInstitutionsContainer } from './app_admin/AdminCustomerInstitutionsContainer';
 import { CentralImportDuplicationCheckPage } from './app_admin/central_import/CentralImportDuplicationCheckPage';
 import { CentralImportPage } from './app_admin/central_import/CentralImportPage';
-import { CentralImportRegistration } from './app_admin/central_import/CentralImportRegistration';
 import { AddEmployeePage } from './institution_admin/AddEmployeePage';
 import { PersonRegisterPage } from './institution_admin/person_register/PersonRegisterPage';
 
@@ -165,12 +164,6 @@ const BasicDataPage = () => {
             exact
             path={UrlPathTemplate.BasicDataCentralImportDuplicateCheck}
             component={CentralImportDuplicationCheckPage}
-            isAuthorized={isAppAdmin}
-          />
-          <PrivateRoute
-            exact
-            path={UrlPathTemplate.BasicDataCentralImportRegistration}
-            component={CentralImportRegistration}
             isAuthorized={isAppAdmin}
           />
           <PrivateRoute
