+++ src/pages/basic_data/app_admin/central_import/CentralImportDuplicationCheckPage.tsx
@@ -1,17 +1,21 @@
+import { LoadingButton } from '@mui/lab';
 import { Box, Button, Paper, Typography } from '@mui/material';
-import { useQuery } from '@tanstack/react-query';
+import { useMutation, useQuery } from '@tanstack/react-query';
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import { useDispatch } from 'react-redux';
 import { Link, useParams } from 'react-router-dom';
+import { createRegistrationFromImportCandidate, fetchImportCandidate } from '../../../../api/registrationApi';
 import { fetchImportCandidates } from '../../../../api/searchApi';
 import { PageSpinner } from '../../../../components/PageSpinner';
 import { StyledPaperHeader } from '../../../../components/PageWithSideMenu';
 import { BackgroundDiv } from '../../../../components/styled/Wrappers';
+import { setNotification } from '../../../../redux/notificationSlice';
 import { SearchResponse } from '../../../../types/common.types';
 import { emptyDuplicateSearchFilter } from '../../../../types/duplicateSearchTypes';
 import { ImportCandidateSummary } from '../../../../types/importCandidate.types';
 import { stringIncludesMathJax, typesetMathJax } from '../../../../utils/mathJaxHelpers';
-import { RegistrationParams, getImportCandidatePagePath } from '../../../../utils/urlPaths';
+import { RegistrationParams, getRegistrationLandingPagePath } from '../../../../utils/urlPaths';
 import NotFound from '../../../errorpages/NotFound';
 import { CentralImportDuplicateSearch } from './CentralImportDuplicateSearch';
 import { CentralImportResultItem } from './CentralImportResultItem';
@@ -19,6 +23,7 @@ import { DuplicateSearchFilterForm } from './DuplicateSearchFilterForm';
 
 export const CentralImportDuplicationCheckPage = () => {
   const { t } = useTranslation();
+  const dispatch = useDispatch();
   const { identifier } = useParams<RegistrationParams>();
   const [duplicateSearchFilters, setDuplicateSearchFilters] = useState(emptyDuplicateSearchFilter);
 
@@ -28,6 +33,28 @@ export const CentralImportDuplicationCheckPage = () => {
     meta: { errorMessage: t('feedback.error.get_registrations') },
   });
   const importCandidate = importCandidateQuery.data?.hits[0];
+
+  const importCandidateMutation = useMutation({
+    mutationFn: async () => {
+      const initialDataRegistration = await fetchImportCandidate(identifier);
+      return await createRegistrationFromImportCandidate(initialDataRegistration);
+    },
+    onSuccess: () => {
+      dispatch(
+        setNotification({
+          message: t('feedback.success.create_registration'),
+          variant: 'success',
+        })
+      );
+    },
+    onError: () =>
+      dispatch(
+        setNotification({
+          message: t('feedback.error.create_registration'),
+          variant: 'error',
+        })
+      ),
+  });
 
   useEffect(() => {
     if (stringIncludesMathJax(importCandidate?.mainTitle)) {
@@ -64,7 +91,6 @@ export const CentralImportDuplicationCheckPage = () => {
               importCandidate={importCandidate}
               setDuplicateSearchFilters={setDuplicateSearchFilters}
             />
-
             <CentralImportDuplicateSearch duplicateSearchFilters={duplicateSearchFilters} />
           </>
         ) : (
@@ -80,9 +106,34 @@ export const CentralImportDuplicationCheckPage = () => {
         </StyledPaperHeader>
 
         <Box sx={{ m: '0.5rem' }}>
-          <Button variant="outlined" color="primary" component={Link} to={getImportCandidatePagePath(identifier)}>
-            {t('basic_data.central_import.import')}
-          </Button>
+          {importCandidate?.importStatus.candidateStatus === 'NOT_IMPORTED' &&
+            (!importCandidateMutation.isSuccess ? (
+              <>
+                <Typography gutterBottom>
+                  {t('basic_data.central_import.create_publication_from_import_candidate')}
+                </Typography>
+
+                <LoadingButton
+                  variant="outlined"
+                  fullWidth
+                  size="small"
+                  color="primary"
+                  disabled={importCandidateMutation.isSuccess}
+                  loading={importCandidateMutation.isLoading}
+                  onClick={() => importCandidateMutation.mutate()}>
+                  {t('basic_data.central_import.create_new')}
+                </LoadingButton>
+              </>
+            ) : (
+              <Button
+                sx={{ mt: '0.5rem' }}
+                variant="contained"
+                color="primary"
+                component={Link}
+                to={getRegistrationLandingPagePath(importCandidateMutation.data.identifier)}>
+                {t('basic_data.central_import.see_publication')}
+              </Button>
+            ))}
         </Box>
       </Paper>
     </Box>
