+++ src/translations/nbTranslations.json
@@ -47,9 +47,10 @@
     "central_import": {
       "author": "$t(registration.contributors.types.Creator)",
       "central_import": "Sentralimport",
+      "create_new": "Opprett ny",
+      "create_publication_from_import_candidate": "Opprett ny publikasjon basert på importpublikasjonen.",
       "duplicate_search_hits": "Søket ga følgende treff:",
       "duplicate_search_no_hits": "Søket ga ingen treff",
-      "import": "Importer",
       "or_search_with": "eller søk med",
       "reset_search_values": "Tilbakestill verdier",
       "search_again": "Søk igjen",
