+++ src/types/importCandidate.types.ts
@@ -13,7 +13,10 @@ import { ReportPublicationInstance } from './publication_types/reportRegistratio
 import { ResearchDataPublicationInstance } from './publication_types/researchDataRegistration.types';
 import { Journal, Publisher, Registration } from './registration.types';
 
-type ImportStatus = 'IMPORTED' | 'NOT_IMPORTED' | 'NOT_APPLICABLE';
+interface ImportStatus {
+  candidateStatus: 'IMPORTED' | 'NOT_IMPORTED' | 'NOT_APPLICABLE';
+  modifiedDate: string;
+}
 
 export interface ImportCandidate extends Omit<Registration, 'type'> {
   type: 'ImportCandidate';
